08-04

+++ Basics/Assets/Code/31-03/I_Enemy.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class I_Enemy : MonoBehaviour, IHitable
 {
-    public void TakeHit(int damage)
+    public void TakeHit(float damage)
     {
         Debug.Log("Enemigo ha sido golpeao y restara vida");
     }
