+++ Basics/Assets/Code/31-03/hitable.cs
@@ -1,5 +1,5 @@
 public interface IHitable
 {
-    void TakeHit(int damage);
+    void TakeHit(float damage);
 
 }
